+++ index.html
@@ -3,58 +3,22 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Test Spotify</title>
+    <link rel="icon" href="./images/undrgrnd_logo.png" type="image/gif" sizes="16x16">
+    <title>UNDR GRND</title>
+    
     <script src="./src/main.js" type="module"></script>  
-    <link rel="stylesheet" type="text/css" href="./styles/style.css" />
-    <style>
-   
-    img{
-        width: 150px;
-        height:150px;
-        border-radius: 50%;
-    }
+    <link rel="stylesheet type="text/css" href="./styles/style.css"/>
     
-    .container{
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-    }
-    
-    .profile{
-        padding: 10px
-    }
-    button{
-        background-color: #09cc4d;
-        font-size: 15px;
-        padding: 19px;
-        margin-top: 25px;
-        color: #ffffff;
-        border-radius: 15px;
-        font-family: Poppins-Regular, sans-serif;
-        box-shadow:
-        0 2.8px 2.2px rgba(0, 0, 0, 0.034),
-        0 6.7px 5.3px rgba(0, 0, 0, 0.048),
-        0 12.5px 10px rgba(0, 0, 0, 0.06),
-        0 22.3px 17.9px rgba(0, 0, 0, 0.072),
-        0 41.8px 33.4px rgba(0, 0, 0, 0.086),
-        0 100px 80px rgba(0, 0, 0, 0.12)
-
-}
-body{
-    background-image: url("./images/video.mp4");
-}
     </style>
 </head>
 <body>
-    <div class ="heading">
-        <img class="logo" src="./images/trans_logo.png" alt="" srcset="" />
-        <video autoplay muted loop id="myVideo">
-            <source src="./images/video.mp4" type="video/mp4">
-        </video>
-        <div id="content"></div>
+    <video autoplay muted loop id="myVideo">
+        <source src="./images/video.mp4" type="video/mp4">
+    </video>
+    <div class="head">
+
     </div>
-    
-    
+    <div id="content"></div>
