--- styles/src/config.js
@@ -1,4 +1,0 @@
-export const clientId = '933790dc3b564fa6b2ed3d8f09ed440f';//the clientId of the app
-export const redirectLogin = "http://localhost:5500/?acceptLogin=true";//needs to the the url you set in spotify just with the ?acceptLogin=true
-export const redirectComplete = "http://localhost:5500/";
-export const scope = 'user-read-email user-follow-modify user-follow-modify user-read-private user-library-read user-library-modify user-top-read'
